📚 Read some books!
This is a squash of a bunch of exploratory commits. I've wanted to have
a bookshelf of sorts on my website for quite a while. The tasks involved
included

- Gathering book reviews from Instagram
- Gathering book reviews from Goodreads
- Gathering book reviews from random Apple Notes
- Building an index page (that was responsive at that!)
- Building a show page for each book.

There's refinements to be made in sharing books that don't have a proper
ISBN, indie published, etc. Also, it feels a bit virtue-signal'ish to be
pointing to all the indie book websites, when in reality I'm still a
person who buys books e-books on Kindle and real books on Amazon. I'm a
work in process.